# microservices/order-management/src/main/openapi/product-management-1.0.yml
openapi: 3.0.1
info:
  title: product-management
  version: "1.0"
paths:
  /pricing:
    post:
      tags:
      - pricing
      operationId: getPrices
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderPricingCreateDTO"
        required: true
      responses:
        "200":
          description: getPrices 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderPricingDTO"
  /products:
    get:
      tags:
      - products
      operationId: list
      parameters:
      - name: page
        in: query
        required: true
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: list 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page_Product_"
    post:
      tags:
      - products
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreateDTO"
        required: true
      responses:
        "200":
          description: create 200 response
          content:
            application/json:
              schema:
                type: object
  /products/{id}:
    get:
      tags:
      - products
      operationId: get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: get 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      tags:
      - products
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreateDTO"
        required: true
      responses:
        "200":
          description: update 200 response
    delete:
      tags:
      - products
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: delete 200 response
  /products/{id}/daily-orders:
    get:
      tags:
      - products
      operationId: getDailyOrders
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        required: true
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: getDailyOrders 200 response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: integer
                  format: int32
  /products/{id}/tags:
    get:
      tags:
      - products
      operationId: listTags
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: listTags 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
  /products/{id}/tags/{tagId}:
    put:
      tags:
      - products
      operationId: addProductTag
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: tagId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: addProductTag 200 response
    delete:
      tags:
      - products
      operationId: removeProductTag
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: tagId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: removeProductTag 200 response
  /tags:
    get:
      tags:
      - tags
      operationId: listTags_1
      parameters:
      - name: page
        in: query
        required: true
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: listTags_1 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page_Tag_"
    post:
      tags:
      - tags
      operationId: createTag
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagCreateDTO"
        required: true
      responses:
        "200":
          description: createTag 200 response
          content:
            application/json:
              schema:
                type: object
  /tags/{id}:
    get:
      tags:
      - tags
      operationId: getTag
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: getTag 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
    put:
      tags:
      - tags
      operationId: updateTag
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagCreateDTO"
        required: true
      responses:
        "200":
          description: updateTag 200 response
    delete:
      tags:
      - tags
      operationId: deleteTag
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: deleteTag 200 response
  /tags/{id}/products:
    get:
      tags:
      - tags
      operationId: listProducts
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: listProducts 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
components:
  schemas:
    OrderItemPricingCreateDTO:
      type: object
      properties:
        productId:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
    OrderItemPricingDTO:
      type: object
      properties:
        productId:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
        unitPrice:
          type: number
    OrderPricingCreateDTO:
      type: object
      properties:
        dateCreated:
          type: string
          format: date
        orderItems:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemPricingCreateDTO"
    OrderPricingDTO:
      type: object
      properties:
        dateCreated:
          type: string
          format: date
        totalAmount:
          type: number
        orderItems:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemPricingDTO"
    Page_Product_:
      type: object
      allOf:
      - $ref: "#/components/schemas/Slice_Product_"
      - properties:
          totalSize:
            type: integer
            format: int64
          totalPages:
            type: integer
            format: int32
    Page_Tag_:
      type: object
      allOf:
      - $ref: "#/components/schemas/Slice_Tag_"
      - properties:
          totalSize:
            type: integer
            format: int64
          totalPages:
            type: integer
            format: int32
    Pageable:
      required:
      - size
      - sort
      type: object
      allOf:
      - $ref: "#/components/schemas/Sort"
      - properties:
          number:
            type: integer
            format: int32
          size:
            type: integer
            format: int32
          mode:
            $ref: "#/components/schemas/Pageable.Mode"
          sort:
            $ref: "#/components/schemas/Sort"
    Pageable.Mode:
      type: string
      enum:
      - CURSOR_NEXT
      - CURSOR_PREVIOUS
      - OFFSET
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        name:
          type: string
        unitPrice:
          type: number
    ProductCreateDTO:
      type: object
      properties:
        unitPrice:
          type: number
        name:
          type: string
    Slice_Product_:
      required:
      - content
      - pageable
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        pageable:
          $ref: "#/components/schemas/Pageable"
        pageNumber:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        empty:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
    Slice_Tag_:
      required:
      - content
      - pageable
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        pageable:
          $ref: "#/components/schemas/Pageable"
        pageNumber:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        empty:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
    Sort:
      required:
      - orderBy
      type: object
      properties:
        orderBy:
          type: array
          items:
            $ref: "#/components/schemas/Sort.Order"
    Sort.Order:
      required:
      - direction
      - ignoreCase
      - property
      type: object
      properties:
        ignoreCase:
          type: boolean
        direction:
          $ref: "#/components/schemas/Sort.Order.Direction"
        property:
          type: string
        ascending:
          type: boolean
    Sort.Order.Direction:
      type: string
      enum:
      - ASC
      - DESC
    Tag:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        name:
          type: string
    TagCreateDTO:
      type: object
      properties:
        name:
          type: string